requirements.txt
python3-nftables
python3-pyroute2
